+++ app/controllers/assessments_controller.rb
@@ -89,10 +89,6 @@ class AssessmentsController < ApplicationController
 
   private
 
-  def parse_date(date_string)
-    date_string.to_s.split('/').last.size == 2 ? Date.strptime(date_string, "%m/%d/%y") : Date.strptime(date_string, "%m/%d/%Y") rescue ""
-  end
-
   def post_params
     params[:assessment] ||= {}