+++ app/controllers/study_types_controller.rb
@@ -88,10 +88,6 @@ class StudyTypesController < ApplicationController
 
   private
 
-  def parse_date(date_string)
-    date_string.to_s.split('/').last.size == 2 ? Date.strptime(date_string, "%m/%d/%y") : Date.strptime(date_string, "%m/%d/%Y") rescue ""
-  end
-
   def post_params
     params[:study_type] ||= {}