+++ app/mailers/exercise_mailer.rb
@@ -1,4 +1,4 @@
-class ExerciseMailer < ActionMailer::Base
+class ExerciseMailer < ApplicationMailer
   default from: "#{DEFAULT_APP_NAME} <#{ActionMailer::Base.smtp_settings[:email]}>"
   add_template_helper(ApplicationHelper)
 
@@ -8,13 +8,5 @@ class ExerciseMailer < ActionMailer::Base
     @exercise = exercise
 
     mail(to: @user.email, subject: @subject + "New Exercise #{@exercise.name} Assigned")
-  end
-
-  protected
-
-  def setup_email
-    @subject = "[#{DEFAULT_APP_NAME.downcase}#{'-development' if Rails.env == 'development'}] "
-    @footer_html = "Change email settings here: <a href=\"#{SITE_URL}/settings\">#{SITE_URL}/settings</a>.<br /><br />".html_safe
-    @footer_txt = "Change email settings here: #{SITE_URL}/settings."
   end
 end