+++ app/mailers/user_mailer.rb
@@ -1,4 +1,4 @@
-class UserMailer < ActionMailer::Base
+class UserMailer < ApplicationMailer
   default from: "#{DEFAULT_APP_NAME} <#{ActionMailer::Base.smtp_settings[:email]}>"
   add_template_helper(ApplicationHelper)
 
@@ -19,11 +19,5 @@ class UserMailer < ActionMailer::Base
 #         reply_to: user.email)
   end
 
-  protected
 
-  def setup_email
-    @subject = "[#{DEFAULT_APP_NAME.downcase}#{'-development' if Rails.env == 'development'}] "
-    @footer_html = "Change email settings here: <a href=\"#{SITE_URL}/settings\">#{SITE_URL}/settings</a>.<br /><br />".html_safe
-    @footer_txt = "Change email settings here: #{SITE_URL}/settings."
-  end
 end