+++ config/initializers/01_version.rb
@@ -3,7 +3,7 @@ module Rely
     MAJOR = 0
     MINOR = 7
     TINY = 2
-    BUILD = "pre"
+    BUILD = nil
     STRING = [MAJOR, MINOR, TINY, BUILD].compact.join('.')
   end
 end