+++ test/integration/study_workflow_test.rb
@@ -1,3 +1,4 @@
+require 'test_helper'
 SimpleCov.command_name "test:integration"
 
 class StudyWorkflowTest < ActionDispatch::IntegrationTest